requirements.txt
flox-lib
pyperclip